Squashed commit of the following:
commit c862c078081a676ee906615f215aeb77cd4b382c
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Tue Jan 3 13:45:07 2023 -0800

    Linter

commit 7650e1a0504dbc3db51fe8273253bc6e51a76fbd
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Tue Jan 3 13:37:11 2023 -0800

    Remove old index.py tests

commit 39b08754112d0beb129e0b764016c8da96a23a87
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Tue Jan 3 13:29:37 2023 -0800

    Remove user library indexing

commit 6395d6ea7ed7c6a112b05210d22ec86aed6d8ed3
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 17:34:33 2022 -0800

    Undo change to monitoring

commit 9ff84c1963c162703e92de504489f4fd0f33d5ea
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 17:28:54 2022 -0800

    Remove index.py and linter

commit 49bb8573c90f23580e03e70cbd476c06ad907dc6
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 17:26:02 2022 -0800

    Remove legacy ursm indexing

commit a72b3140dae3316cda8b0d8dab09ca0a74656d6a
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 17:02:48 2022 -0800

    Remove legacy social feature indexing

commit 7e3b19e490a4a71a9234f918c53c10ad33ec2cc4
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 16:56:39 2022 -0800

    Remove legacy user indexing

commit 66e574ea1fa514b48c64d11a1f09612ccd96b405
Author: Michelle Brier <michelle.brier4@gmail.com>
Date:   Thu Dec 8 16:38:37 2022 -0800

    Remove legacy track + playlist indexing code
